+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Query filters can now be overloadable.
+
 ### Fixed
 - Documentation was incorrectly calling the wrong method #156
 - Incorrect timezone documentation for SC #145